Resolve issues opening passive data connection  with a non-local IP NATed
When Waarp Gateway FTP is behind a natted IP, the <serveraddr> tag in the config file
is used to configure the externally visible IP.
However, if serveraddr is given, Waarp Gateway FTP tries to open a socket on this specific
address to receive a passive connection.
When the natted IP given in serveraddr is not local, the socket cannot be opened.

This fix does the following :

- The passive socket is always opened on the same IP as the authentication socket
- The IP address given in serveraddr is only considered when generating the reply
  to the PASV command